file-agent/requirements.txt
grpcio>=1.74.0
grpcio-tools>=1.74.0
python-dotenv>=1.0.0
